requirements.txt
numpy==1.24.3
Pillow==9.5.0
six==1.16.0
wxPython==4.2.0
